core 파트 완성

+++ SpaceshipGame/TextManagerClass.h
@@ -12,7 +12,7 @@ namespace Text
 
 		ITextClass* GetTextObject(int idx) const;
 		inline UINT GetFontMask() const { return m_FontMask; }
-		inline UINT GetTextCount() const { return m_UITexts.size(); }
+		inline UINT GetTextCount() const { return static_cast<UINT>(m_UITexts.size()); }
 
 		void Load(ID id, const std::wstring& text, Graphic::Font::ID font = Graphic::Font::ID::DEFAULT);
 		void Release();
